pyproject.toml
[tool.poetry]
name = "maeser"
version = "0.1.0"
description = "A package for building RAG chatbot applications for educational contexts."
authors = ["Carson Bush <hyperdriveguy@byui.edu>", "Blaine Freestone <blainefreestone@gmail.com>"]
include = [
    "maeser/data/*",
    "maeser/data/templates/*",
    "maeser/data/static/*"
]

# Dependencies required by your project
[tool.poetry.dependencies]
# Specify where your Python interpreter is located
# Replace with your Python version if different
python = "<4.0,>=3.10"

# Langchain libraries for text processing and embeddings
langchain = "^0.2.8"
langchain_core = "^0.2.19"
langchain_community = "^0.2.7"
langchain_openai = "^0.1.16"
langchain-text-splitters = "^0.2.2"
langgraph = "^0.1.8"

# FAISS for vector stores (CPU version by default, GPU as optional)
# pip install maeser[gpu] for the GPU variant
faiss-cpu = "^1.8.0.post1"
# faiss-gpu = { version = "^1.7.2", optional = true }

# Web backend
flask = "^3.0.3"
flask-login = "^0.6.3"

# Turning markdown into HTML for display
markdown = "^3.6"
pymdown-extensions = "^10.8.1"
markdownify = "^0.13.1"

# Unit testing
ragas = "^0.1.10"

# Configuration
pyYAML = "^6.0.1"

# For HTTP requests
requests = "^2.32.3"

# Declare your project's development dependencies here
[tool.poetry.dev-dependencies]
pytest = "^8.2"

# Documentation
sphinx_rtd_theme = "^2.0.0"
myst_parser = "^3.0.1"
sphinx-book-theme = "^1.1.3"

# Build system configuration
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Define extra dependencies
[tool.poetry.extras]
gpu = ["faiss-gpu"]
